Update version to 1.2.0
