+++ README_CN.md
@@ -29,6 +29,7 @@ SQLlin 支持如下平台：
 
 - Multiplatform Common
 - Android (6.0+)
+- JVM (Java 11+, since `1.2.0`)
 - iOS (x64, arm64, simulatorArm64)
 - macOS (x64, arm64)
 - watchOS (x64, arm32, arm64, simulatorArm64, deviceArm64)
