+++ gradle.properties
@@ -1,4 +1,4 @@
-VERSION=1.1.1
+VERSION=1.2.0
 GROUP=com.ctrip.kotlin
 
 kotlinVersion=1.9.0
