+++ sqllin-driver/src/jvmMain/kotlin/com/ctrip/sqllin/driver/ConcurrentDatabaseConnection.kt
@@ -60,12 +60,6 @@ internal class ConcurrentDatabaseConnection(private val delegateConnection: Data
         delegateConnection.close()
     }
 
-    @Deprecated(
-        message = "The property closed has been deprecated, please use the isClosed to replace it",
-        replaceWith = ReplaceWith("isClosed")
-    )
-    override val closed: Boolean
-        get() = delegateConnection.isClosed
     override val isClosed: Boolean
         get() = delegateConnection.isClosed
 }
