+++ sqllin-driver/src/nativeMain/kotlin/com/ctrip/sqllin/driver/ConcurrentDatabaseConnection.kt
@@ -66,12 +66,6 @@ internal class ConcurrentDatabaseConnection(
         accessLock.close()
     }
 
-    @Deprecated(
-        message = "The property closed has been deprecated, please use the isClosed to replace it",
-        replaceWith = ReplaceWith("isClosed")
-    )
-    override val closed: Boolean
-        get() = delegateConnection.isClosed
     override val isClosed: Boolean
         get() = delegateConnection.isClosed
