+++ sqllin-driver/src/nativeMain/kotlin/com/ctrip/sqllin/driver/ExtensionNative.kt
@@ -18,7 +18,6 @@ package com.ctrip.sqllin.driver
 
 import com.ctrip.sqllin.driver.cinterop.NativeDatabase.Companion.openNativeDatabase
 import com.ctrip.sqllin.driver.platform.Lock
-import com.ctrip.sqllin.driver.platform.separatorChar
 import com.ctrip.sqllin.driver.platform.withLock
 import platform.posix.remove
 
@@ -27,9 +26,7 @@ import platform.posix.remove
  * @author yaqiao
  */
 
-public fun String.toDatabasePath(): DatabasePath = NativeDatabasePath(this)
-
-internal value class NativeDatabasePath(val pathString: String) : DatabasePath
+public fun String.toDatabasePath(): DatabasePath = StringDatabasePath(this)
 
 private val connectionCreationLock = Lock()
 public actual fun openDatabase(config: DatabaseConfiguration): DatabaseConnection = connectionCreationLock.withLock {
@@ -52,70 +49,8 @@ public actual fun openDatabase(config: DatabaseConfiguration): DatabaseConnectio
     if (config.isReadOnly) realConnection else ConcurrentDatabaseConnection(realConnection)
 }
 
-private fun DatabaseConfiguration.diskOrMemoryPath(): String =
-    if (inMemory) {
-        if (name.isBlank())
-            ":memory:"
-        else
-            "file:$name?mode=memory&cache=shared"
-    } else {
-        require(name.isNotBlank()) { "Database name cannot be blank" }
-        getDatabaseFullPath((path as NativeDatabasePath).pathString, name)
-    }
-
-private fun getDatabaseFullPath(dirPath: String, name: String): String {
-    val param = when {
-        dirPath.isEmpty() -> name
-        name.isEmpty() -> dirPath
-        else -> join(dirPath, name)
-    }
-    return fixSlashes(param)
-}
-
-private fun join(prefix: String, suffix: String): String {
-    val haveSlash = (prefix.isNotEmpty() && prefix.last() == separatorChar)
-            || (suffix.isNotEmpty() && suffix.first() == separatorChar)
-    return buildString {
-        append(prefix)
-        if (!haveSlash)
-            append(separatorChar)
-        append(suffix)
-    }
-}
-
-private fun fixSlashes(origPath: String): String {
-    // Remove duplicate adjacent slashes.
-    var lastWasSlash = false
-    val newPath = origPath.toCharArray()
-    val length = newPath.size
-    var newLength = 0
-    val initialIndex = if (origPath.startsWith("file://", true)) 7 else 0
-    for (i in initialIndex ..< length) {
-        val ch = newPath[i]
-        if (ch == separatorChar) {
-            if (!lastWasSlash) {
-                newPath[newLength++] = separatorChar
-                lastWasSlash = true
-            }
-        } else {
-            newPath[newLength++] = ch
-            lastWasSlash = false
-        }
-    }
-    // Remove any trailing slash (unless this is the root of the file system).
-    if (lastWasSlash && newLength > 1) {
-        newLength--
-    }
-
-    // Reuse the original string if possible.
-    return if (newLength != length) buildString(newLength) {
-        append(newPath)
-        setLength(newLength)
-    } else origPath
-}
-
 public actual fun deleteDatabase(path: DatabasePath, name: String): Boolean {
-    val baseName = getDatabaseFullPath((path as NativeDatabasePath).pathString, name)
+    val baseName = getDatabaseFullPath((path as StringDatabasePath).pathString, name)
     remove("$baseName-shm")
     remove("$baseName-wal")
     remove("$baseName-journal")
